Improve readability
Break up html elements to attempt to not cross the 80-character line limit. This helps improve readability.

The 80-character line limit originated with the IBM punch card, which had 80 columns. From https://en.wikipedia.org/wiki/Punched_card: 'A punched card or punch card is a piece of stiff paper that can be used to contain digital information represented by the presence or absence of holes in predefined positions. The information might be data for data processing applications or, in earlier examples, used to directly control automated machinery.'