Create a function for 'thrilling-heroics'
This extracts all of the crazy code we had stuffed into onclick in our
first gif set. The code was difficult to read as it was all one big
line. As a function, it’s much easier to split the code up and see
what’s going on. Big win!